+++ litemall-admin/src/views/screen/screenDevice.vue
@@ -287,20 +287,23 @@ export default {
       })
     },
     handleDelete(row) {
-      deleteScreenDevice(row.id)
-        .then(response => {
-          this.$notify.success({
-            title: '成功',
-            message: '删除成功'
+      this.$confirm('确定删除吗？').then(_ => {
+        deleteScreenDevice(row.id)
+          .then(response => {
+            this.$notify.success({
+              title: '成功',
+              message: '删除成功'
+            })
+            this.getList()
           })
-          this.getList()
-        })
-        .catch(response => {
-          this.$notify.error({
-            title: '失败',
-            message: response.data.errmsg
+          .catch(response => {
+            this.$notify.error({
+              title: '失败',
+              message: response.data.errmsg
+            })
           })
-        })
+      }).catch(_ => {
+      })
     },
     handleReboot(row) {
       reboot(row.cardId)
