+++ litemall-admin/src/views/screen/source.vue
@@ -409,20 +409,23 @@ export default {
       })
     },
     handleDelete(row) {
-      deleteSource(row.sourceId)
-        .then(response => {
-          this.$notify.success({
-            title: '成功',
-            message: '删除成功'
+      this.$confirm('确定删除吗？').then(_ => {
+        deleteSource(row.sourceId)
+          .then(response => {
+            this.$notify.success({
+              title: '成功',
+              message: '删除成功'
+            })
+            this.getList()
           })
-          this.getList()
-        })
-        .catch(response => {
-          this.$notify.error({
-            title: '失败',
-            message: response.data.errmsg
+          .catch(response => {
+            this.$notify.error({
+              title: '失败',
+              message: response.data.errmsg
+            })
           })
-        })
+      }).catch(_ => {
+      })
     }
   }
 }
